ZA: Move the person detail class in views.py to its own file
The logger global variable and the AttendanceAPIDown exception
where moved to this file as well, since they are only used here.
