+++ pombola/south_africa/views/__init__.py
@@ -2,5 +2,5 @@ from .constants import *
 from .geolocalization import *
 from .home import *
 from .organisations import *
+from .person import *
 from .views import *
-
